SCB-1124 Rename the Handler name of Omega transport
