+++ omega/omega-transport/omega-transport-servicecomb/src/test/java/org/apache/servicecomb/pack/omega/transport/servicecomb/PackProviderHandlerTest.java
@@ -17,11 +17,16 @@
 
 package org.apache.servicecomb.pack.omega.transport.servicecomb;
 
-import static org.hamcrest.core.Is.is;
+import static org.apache.servicecomb.pack.omega.context.OmegaContext.GLOBAL_TX_ID_KEY;
+import static org.apache.servicecomb.pack.omega.context.OmegaContext.LOCAL_TX_ID_KEY;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
@@ -33,14 +38,14 @@ import org.apache.servicecomb.swagger.invocation.AsyncResponse;
 import org.junit.Before;
 import org.junit.Test;
 
-public class SagaConsumerHandlerTest {
+import org.mockito.invocation.InvocationOnMock;
+import org.mockito.stubbing.Answer;
+
+public class PackProviderHandlerTest {
 
   private static final String globalTxId = UUID.randomUUID().toString();
 
   private static final String localTxId = UUID.randomUUID().toString();
-
-  @SuppressWarnings("unchecked")
-  private final IdGenerator<String> idGenerator = mock(IdGenerator.class);
 
   private final OmegaContext omegaContext = new OmegaContext(new IdGenerator<String>() {
 
@@ -54,34 +59,45 @@ public class SagaConsumerHandlerTest {
 
   private final AsyncResponse asyncResponse = mock(AsyncResponse.class);
 
-  private final SagaConsumerHandler handler = new SagaConsumerHandler(omegaContext);
+  private final PackProviderHandler handler = new PackProviderHandler(omegaContext);
 
   @Before
   public void setUp() {
-    when(idGenerator.nextId()).thenReturn(globalTxId, localTxId);
+    omegaContext.clear();
   }
 
   @Test
-  public void keepHeaderUnchangedIfContextAbsent() throws Exception {
+  public void setUpOmegaContextInTransactionRequest() throws Exception {
     Map<String, String> context = new HashMap<>();
+    context.put(GLOBAL_TX_ID_KEY, globalTxId);
+    context.put(LOCAL_TX_ID_KEY, localTxId);
     when(invocation.getContext()).thenReturn(context);
+
+    doAnswer(new Answer<Void>() {
+      // Just verify the context setting
+      @Override
+      public Void answer(InvocationOnMock invocation) throws Throwable {
+        assertThat(omegaContext.globalTxId(), is(globalTxId));
+        assertThat(omegaContext.localTxId(), is(localTxId));
+        return null;
+      }
+    }).when(invocation).next(asyncResponse);
+
 
     handler.handle(invocation, asyncResponse);
 
-    assertThat(invocation.getContext().isEmpty(), is(true));
+    // Now the context should be clean
+    assertThat(omegaContext.globalTxId(), is(nullValue()));
+    assertThat(omegaContext.localTxId(), is(nullValue()));
   }
 
   @Test
-  public void interceptTransactionIdInHeaderIfContextPresent() throws Exception {
-    omegaContext.setGlobalTxId(globalTxId);
-    omegaContext.setLocalTxId(localTxId);
-
-    Map<String, String> context = new HashMap<>();
-    when(invocation.getContext()).thenReturn(context);
+  public void doNothingInNonTransactionRequest() throws Exception {
+    when(invocation.getContext()).thenReturn(Collections.EMPTY_MAP);
 
     handler.handle(invocation, asyncResponse);
 
-    assertThat(invocation.getContext().get(OmegaContext.GLOBAL_TX_ID_KEY), is(globalTxId));
-    assertThat(invocation.getContext().get(OmegaContext.LOCAL_TX_ID_KEY), is(localTxId));
+    assertThat(omegaContext.globalTxId(), is(nullValue()));
+    assertThat(omegaContext.localTxId(), is(nullValue()));
   }
 }
